requirements.txt
kafka-python
aiohttp
psycopg2
